Refactoring the LTSpiceLogReader dividing it into two separate classes. The base class supports all statistical and export functions, the LTSpiceLogReader contains only the log parser
prepare_testbench() method now only accepts *args and **kwargs

A working Worst Case Example.
