+++ PyLTSpice/sim/tookit/sim_analysis.py
@@ -23,77 +23,11 @@ from collections import OrderedDict
 from functools import wraps
 from typing import Union, Optional
 
+from log.logfile_data import LogfileData
 from ...sim.sim_runner import AnyRunner
 from ...editor.base_editor import BaseEditor
 from ...sim.simulator import Simulator
 from ...log.ltsteps import LTSpiceLogReader
-
-
-class LogFileAnalysis(LTSpiceLogReader):
-    """
-    This is a subclass of LTSpiceLogReader that is used to analyse the log file of a simulation.
-    The super class constructor is bypassed and only their attributes are initialized
-    """
-    def __init__(self, stepset, dataset, encoding):
-        self.stepset = stepset
-        self.dataset = dataset
-        self.logname = None
-        self.encoding = encoding
-        self.step_count = len(stepset)
-        self.measure_count = len(dataset)
-
-    def plot_histogram(self, param, bins=50, normalized=True, sigma=3.0, title=None, image_file=None, **kwargs):
-        """
-        Plots a histogram of the parameter
-        """
-        import numpy as np
-        from scipy.stats import norm
-        import matplotlib.pyplot as plt
-
-        x = np.array(self.dataset[param], dtype=float)
-        mu = x.mean()
-        mn = x.min()
-        mx = x.max()
-        sd = np.std(x)
-        sigmin = mu - sigma*sd
-        sigmax = mu + sigma*sd
-
-        # Automatic calculation of the range
-        axisXmin = mu - (sigma + 1) * sd
-        axisXmax = mu + (sigma + 1) * sd
-
-        if mn < axisXmin:
-            axisXmin = mn
-
-        if mx > axisXmax:
-            axisXmax = mx
-
-        n, bins, patches = plt.hist(x, bins, density=normalized, facecolor='green', alpha=0.75,
-                                    range=(axisXmin, axisXmax))
-        axisYmax = n.max() * 1.1
-
-        if normalized:
-            # add a 'best fit' line
-            y = norm.pdf(bins, mu, sd)
-            l = plt.plot(bins, y, 'r--', linewidth=1)
-            plt.axvspan(mu - sigma*sd, mu + sigma*sd, alpha=0.2, color="cyan")
-            plt.ylabel('Distribution [Normalised]')
-        else:
-            plt.ylabel('Distribution')
-        plt.xlabel(param)
-
-        if title is None:
-            fmt = '%g'
-            title = (r'$\mathrm{Histogram\ of\ %s:}\ \mu='+fmt+r',\ stdev='+fmt+r',\ \sigma=%d$') % (param, mu, sd, sigma)
-
-        plt.title(title)
-
-        plt.axis([axisXmin, axisXmax, 0, axisYmax ])
-        plt.grid(True)
-        if image_file is not None:
-            plt.savefig(image_file)
-        else:
-            plt.show()
 
 
 class SimAnalysis(object):
@@ -151,7 +85,7 @@ class SimAnalysis(object):
     def __getitem__(self, item):
         return self.simulations[item]
 
-    def read_logfiles(self) -> LogFileAnalysis:
+    def read_logfiles(self) -> LogfileData:
         """Reads the log files and returns a dictionary with the results"""
         all_stepset = {}
         all_dataset = OrderedDict()
@@ -170,5 +104,5 @@ class SimAnalysis(object):
                 else:
                     all_dataset[param].extend(log_results.dataset[param])
         # Now reusing the last log_results object to store the results
-        return LogFileAnalysis(all_stepset, all_dataset, log_results.encoding)
+        return LogfileData(all_stepset, all_dataset)
